chore(theme): rename files
