+++ pages/_document.tsx
@@ -1,7 +1,7 @@
 import { ColorModeScript } from "@chakra-ui/react";
 import NextDocument, { Html, Head, Main, NextScript } from "next/document";
 
-import theme from "../theme/index";
+import { theme } from "../theme";
 
 //boilerplate for persistent mode when reload page
 //not working correctly..
